rent_telegram_bot/requirements.txt
asgiref==3.6.0
behave==1.2.6
cachier==1.5.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==7.1
cryptography==40.0.2
Django==4.2
django-extensions==3.2.1
django-filter==23.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
loguru==0.5.3
markup==0.2
MarkupSafe==2.1.2
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pathtools==0.1.2
Pillow==9.5.0
plumbum==1.7.1
portalocker==2.7.0
pycparser==2.21
pymongo==3.12.1
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyTelegramBotAPI==4.10.0
python-rex==0.4
pytz==2023.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.3
style==1.1.0
telebot==0.0.5
update==0.0.1
updater==0.3.2
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.2.3
